+++ app/components/widgets/cloud-widget/content/cloud-widget-content.ts
@@ -17,7 +17,7 @@ export class CloudWidgetContentComponent implements OnChanges, OnInit {
   @Input() showWorks: boolean = false;
   @Input() reload: boolean;
   @Input() collapsed: boolean;
-  @Output() collapsedChange: EventEmitter<boolean> = new EventEmitter();
+  @Output() collapsedChange: EventEmitter<any> = new EventEmitter();
 
   cloud: any = {};
   loading: boolean;
@@ -87,7 +87,10 @@ export class CloudWidgetContentComponent implements OnChanges, OnInit {
     let handlerSoft = () => {
       this.cloudWidgetService.rebootInstance(this.serviceName, id, 'soft')
         .then(
-          () => this.nav.present(this.toastService.success('Redémarrage en cours...')),
+          () => {
+            this.nav.present(this.toastService.success('Redémarrage en cours...'));
+            this.getInfos();
+          },
           (err) => this.nav.present(this.toastService.success(`Une erreur est survenue (${JSON.stringify(err)})`))
         );
     };