+++ app/components/widgets/dedicated-widget/content/dedicated-widget-content.ts
@@ -16,7 +16,7 @@ export class DedicatedWidgetContentComponent implements OnChanges, OnInit {
   @Input() reload: boolean;
   @Input() collapsed: boolean;
   @Input() showWorks: boolean = false;
-  @Output() collapsedChange: EventEmitter<boolean> = new EventEmitter();
+  @Output() collapsedChange: EventEmitter<any> = new EventEmitter();
   viewMode: string = 'general';
   loading: boolean;
   tasksLoaded: boolean = false;