+++ app/components/widgets/private-database-widget/content/private-database-widget-content.ts
@@ -16,7 +16,7 @@ export class PrivateDatabaseWidgetContentComponent implements OnChanges, OnInit 
   @Input() showWorks: boolean = false;
   @Input() reload: boolean;
   @Input() collapsed: boolean;
-  @Output() collapsedChange: EventEmitter<boolean> = new EventEmitter();
+  @Output() collapsedChange: EventEmitter<any> = new EventEmitter();
 
   bdd: any = {};
   loading: boolean;